+++ lib/pairings_engine/tournament_export.ex
@@ -39,7 +39,7 @@ defmodule PairingsEngine.TournamentExport do
   @tournament_fields ~w(
     name type venue city federation start_date end_date organizer
     chief_arbiter deputy_arbiter time_control rounds_count
-    points_win points_draw points_loss bye_value presence_value abs_value
+    points_win points_draw points_loss bye_value requested_bye_type presence_value abs_value
     abs_jusque abs_nbfois absent_counts_as_vur
     presence_on_allocated_bye tiebreaks acceleration
     status standard rate_of_play organizer_club_number round_dates